+++ src/commands/generate/__tests__/generatorTest.ts
@@ -28,11 +28,11 @@ import {
 import * as fs from "fs-extra";
 import * as path from "path";
 import { directory } from "tempy";
+import { ITypeGenerationFlags } from "../../../types/typeGenerationFlags";
+import { isFlavorizable } from "../../../utils/flavorizingUtils";
 import { loadConjureDefinition } from "../generateCommand";
 import { generate } from "../generator";
 import { typeNameToFilePath } from "../simpleAst";
-import { ITypeGenerationFlags } from "../typeGenerationFlags";
-import { isFlavorizable } from "../utils";
 import {
     DEFAULT_TYPE_GENERATION_FLAGS,
     FLAVORED_TYPE_GENERATION_FLAGS,
