+++ src/commands/generate/__tests__/simpleAstTest.ts
@@ -19,11 +19,11 @@ import { ErrorCode, HttpMethod, PrimitiveType } from "conjure-api";
 import * as fs from "fs";
 import * as path from "path";
 import { directory } from "tempy";
+import { createHashableTypeName } from "../../../utils/hashingUtils";
 import { generateError } from "../errorGenerator";
 import { generateService } from "../serviceGenerator";
 import { SimpleAst } from "../simpleAst";
 import { generateEnum } from "../typeGenerator";
-import { createHashableTypeName } from "../utils";
 import { DEFAULT_TYPE_GENERATION_FLAGS } from "./resources/constants";
 
 describe("simpleAst", () => {
