+++ src/utils/flavorizingUtils.ts
@@ -1,0 +1,29 @@
+/**
+ * @license
+ * Copyright 2025 Palantir Technologies, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+import { IType, PrimitiveType } from "conjure-api";
+
+const NON_FLAVORIZABLE_TYPES = new Set<PrimitiveType>([
+    PrimitiveType.ANY,
+    PrimitiveType.BOOLEAN,
+    PrimitiveType.BINARY,
+    PrimitiveType.DATETIME,
+]);
+
+export function isFlavorizable(type: IType, flavorizedAliases: boolean): boolean {
+    return flavorizedAliases && IType.isPrimitive(type) && !NON_FLAVORIZABLE_TYPES.has(type.primitive);
+}
